Bug 2013107 - Separate first session ping with attribution ping. Implement Playstore attribution retry mechanism.  a=pascalc DONTBUILD
Original Revision: https://phabricator.services.mozilla.com/D282450

Differential Revision: https://phabricator.services.mozilla.com/D291367
